Combine message until funcs - compareMessagesForGrouping - isSameGroup - getMessageCreatedAt
